WPA2 Downgrade-Attack to force use of RC4
This is a Downgrade-Attack based on the mc-mitm.py implementation. Here an Access Point (AP) which supports CCMP and TKIP will be cloned on another channel only supporting TKIP. All Beacons from the real AP will be modified so that only TKIP is advertised. This tool tests whether a device connects to the AP using TKIP and performs the first three messages of the 4-way handshake. If the attack succeeds the Group Temporal Key (GTK) will be encrypted with RC4 during the third handshake message. This attack is based on the research paper "Predicting, Decrypting, and Abusing WPA2/802.11 Group Keys" from Mathy Vanhoef and Frank Piessens.