Tool to test which SAE groups a device chooses
This tool tests which SAE Groups a device chooses while connecting to a WPA3-Access Point (AP). A WPA3-AP is being setup using hostapd. The AP rejects all SAE groups a device chooses while connecting to test if and which alternative groups it chooses. The chosen groups can be used for potential Downgrade- or Denial-of-Service Attacks. The idea is based on the research paper "Dragonblood: Analyzing the Dragonfly Handshake of WPA3 and EAP-pwd" from Mathy Vanhoef and Eyal Ronen.