Downgrade-Attack for WPA3-Transition-Mode APs
This is a downgrade-attack to test whether a device automatically connects to a WPA2-Access Point (AP). The SSID has to be set to the SSID of a known WPA3-Transition-Mode AP. If the attack succeeds the handshake messages will be saved to a .pcap file for potential brute-force attacks.